api/services/statics/requirements.txt
pytest
fastapi
python-magic
filetype
pytest-asyncio
httpx
uvicorn
python-multipart